Final changes to the website ui
