+++ my-app/src/App.js
@@ -1,92 +1,15 @@
 import React, { useState } from "react";
-import ReactDOM from "react-dom";
+import Login from "./Login";
 import "./style.css";
 
+//This is the entry page to get into the app
 function App() {
-  // React States
-  const [errorMessages, setErrorMessages] = useState({});
-  const [isSubmitted, setIsSubmitted] = useState(false);
-
-  // User Login info
-  const database = [
-    {
-      username: "user1",
-      password: "pass1"
-    },
-    {
-      username: "user2",
-      password: "pass2"
-    }
-  ];
-
-  const errors = {
-    uname: "invalid username",
-    pass: "invalid password"
-  };
-
-  const handleSubmit = (event) => {
-    //Prevent page reload
-    event.preventDefault();
-
-    var { uname, pass } = document.forms[0];
-
-    // Find user login info
-    const userData = database.find((user) => user.username === uname.value);
-
-    // Compare user info
-    if (userData) {
-      if (userData.password !== pass.value) {
-        // Invalid password
-        setErrorMessages({ name: "pass", message: errors.pass });
-      } else {
-        setIsSubmitted(true);
-      }
-    } else {
-      // Username not found
-      setErrorMessages({ name: "uname", message: errors.uname });
-    }
-  };
-
-  // Generate JSX code for error message
-  const renderErrorMessage = (name) =>
-    name === errorMessages.name && (
-      <div className="error">{errorMessages.message}</div>
-    );
-
-  // JSX code for login form
-  const renderForm = (
-    <div className="form">
-      <form onSubmit={handleSubmit}>
-        <div className="input-container" >
-          <label>Email or Username </label>
-          <input type="text" name="uname" placeholder="Enter your email or username" required />
-          {renderErrorMessage("uname")}
-        </div>
-        <div className="input-container">
-          <label>Password </label>
-          <input type="password" name="pass" placeholder="Enter your password" required />
-          {renderErrorMessage("pass")}
-        </div>
-        <div className="button-container">
-          <input type="Submit" value="Login now" />
-        </div>
-        <div>
-        
-        </div>
-      </form>
-    </div>
-  );
-
   return (
     <div className="app">
-      <div className="login-form">
-        <div className="title">Log into your account</div>
-        {isSubmitted ? <div>User is successfully logged in</div> : renderForm}
-      </div>
+      <Login />
+      
     </div>
-  );
+  )
 }
 
-const rootElement = document.getElementById("root");
-ReactDOM.render(<App />, rootElement);
-export default App;
+export default App
