+++ my-app/src/Feed.css
@@ -1,0 +1,8 @@
+.feed{
+    width: 80%;
+    display: flex;
+    flex-direction: row;
+    margin: 0 auto;
+    margin-top: 20px;
+    justify-content: space-between;
+}
