+++ my-app/src/Login.js
@@ -1,0 +1,97 @@
+import React, {useState} from "react";
+import Blog from "./Blog";
+import "./style.css";
+//After verifying and logging in we get into the blog
+function Login() {
+    // React States
+  const [errorMessages, setErrorMessages] = useState({});
+  const [isSubmitted, setIsSubmitted] = useState(false);
+
+  // User Login info
+  const database = [
+    {
+      username: "user1",
+      password: "pass1"
+    },
+    {
+      username: "user2",
+      password: "pass2"
+    }
+  ];
+
+  const errors = {
+    uname: "invalid username",
+    pass: "invalid password"
+  };
+
+  const handleSubmit = (event) => {
+    //Prevent page reload
+    event.preventDefault();
+
+    var { uname, pass } = document.forms[0];
+
+    // Find user login info
+    const userData = database.find((user) => user.username === uname.value);
+
+    // Compare user info
+    if (userData) {
+      if (userData.password !== pass.value) {
+        // Invalid password
+        setErrorMessages({ name: "pass", message: errors.pass });
+      } else {
+        if(setIsSubmitted(true)){
+            return(
+                <Blog />
+            )
+        }
+      }
+    } else {
+      // Username not found
+      setErrorMessages({ name: "uname", message: errors.uname });
+    }
+  };
+
+  // Generate JSX code for error message
+  const renderErrorMessage = (name) =>
+    name === errorMessages.name && (
+      <div className="error">{errorMessages.message}</div>
+    );
+
+  // JSX code for login form
+  const renderForm = (
+    <div className="form">
+      <form onSubmit={handleSubmit}>
+        <div className="input-container" >
+          <label>Email or Username </label>
+          <input type="text" name="uname" placeholder="Enter your email or username" required />
+          {renderErrorMessage("uname")}
+        </div>
+        <div className="input-container">
+          <label>Password </label>
+          <input type="password" name="pass" placeholder="Enter your password" required />
+          {renderErrorMessage("pass")}
+        </div>
+        <div className="button-container">
+          <input type="Submit" value="Login now" />
+        </div>
+        <div>
+        
+        </div>
+      </form>
+    </div>
+  );
+
+    
+    return (
+    <div className="login">
+      <div className="login-form">
+        {isSubmitted ? <Blog /> : renderForm}
+      </div>
+    </div>
+    );
+    
+}
+
+//const rootElement = document.getElementById("root");
+//ReactDOM.render(<App />, rootElement);
+export default Login
