+++ my-app/src/Post.css
@@ -1,0 +1,9 @@
+.post{
+    flex: 1;
+    padding: 30px 150px;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+
+}
