+++ my-app/src/Welcome.css
@@ -1,0 +1,6 @@
+.welcome{
+    background-position: center center;
+    background-size: cover;
+    height: 60 px;
+}
+
