+++ my-app/src/index.css
@@ -12,11 +12,4 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
-select.form-control,
-textarea.form-control,
-input.form-control {
-  font-size: 1rem;
-}
-input[type=file] {
-  width: 100%;
-}
+
